+++ GeoGL/ellipsoid.h
@@ -13,7 +13,7 @@ public:
 	Ellipsoid(double semimajor, double semiminor, double semiminor2);
 	~Ellipsoid(void);
 	glm::vec3 toVector(double lon, double lat);
-	double heightAboveSurfaceAtPoint(glm::vec3 P);
+	double heightAboveSurfaceAtPoint(glm::dvec3 P);
 	double A() { return a; }
 	double B() { return b; }
 	double C() { return c; }
